Retire rust size baselines
