+++ .pm/tasks/task_d2e428f00e5047e581061c8cb75963ef.execution.md
@@ -1,7 +1,7 @@
 # task_d2e428f00e5047e581061c8cb75963ef Execution Log
 
 - task_uid: task_d2e428f00e5047e581061c8cb75963ef
-- title: clear rust structural slicing baseline
+- title: clear rust size baselines
 - owner_role: runtime_engineer
 - worktree_hint: /home/scc/worktrees/oasis7-engineering-clear-rust-structural-slicing-baseline
 
@@ -15,3 +15,7 @@ Example:
 ## 2026-04-20 00:28:00 CST / runtime_engineer
 - 完成内容: 将 `crates/oasis7`、`crates/oasis7_consensus`、`crates/oasis7_node` 内所有 `split_part` / `impl_part` 存量文件改成语义化文件名，清空结构切片扫描结果；同时把 `tests_conversation_flow.rs`、`tests_consensus_signatures.rs`、`tests_storage_replication.rs` 拆到非模式命名的子文件，退役 `doc/.governance/rust-structural-slicing-baseline.tsv`，并将 `scripts/check-rust-file-size.sh` 改为结构切片实时归零门禁。
 - 遗留事项: 仍有 5 个生产 Rust 超限文件和 3 个测试超限文件留在 `rust-oversized-file-baseline.tsv`，后续继续按 burn-down 任务处理。
+
+## 2026-04-20 13:23:27 CST / runtime_engineer
+- 完成内容: 继续在同一任务内把 `control_plane.rs`、`llm_sidecar.rs`、`launcher_core.rs`、`replication.rs`、`auth_actions.rs`、`main_tests.rs`、`crates/oasis7_net/src/tests.rs` 拆到语义化子模块，令当前 Rust 超限扫描与结构切片扫描都归零；同步删除 `doc/.governance/rust-oversized-file-baseline.tsv`，并把 `scripts/check-rust-file-size.sh` 收口为不依赖任何 frozen baseline 的实时零扫描门禁。
+- 遗留事项: 无；后续若再次出现任一超限 Rust 文件或 `split_part/include!` 结构切片，required gate 直接阻断。
