+++ crates/oasis7/src/viewer/runtime_live/tests/auth_actions.rs
@@ -3,6 +3,9 @@ use std::collections::BTreeMap;
 use std::io::Read;
 use std::sync::{Arc, Mutex};
 
+#[path = "auth_actions_authoritative_recovery.rs"]
+mod authoritative_recovery_tests;
+
 #[test]
 fn runtime_agent_chat_script_mode_requires_llm_mode() {
     let mut server = ViewerRuntimeLiveServer::new(
@@ -1026,310 +1029,4 @@ fn runtime_agent_chat_echo_removed_old_brand_env_is_ignored() {
         crate::simulator::WorldEventKind::AgentSpoke { agent_id: event_agent_id, message, .. }
             if event_agent_id == &agent_id && message == "[qa-echo] hello removed old brand runtime echo"
     )));
-}
-
-#[test]
-fn runtime_agent_chat_rejects_intent_seq_conflict_on_payload_change() {
-    let _guard = lock_test_llm_env();
-    let mut server = ViewerRuntimeLiveServer::new(
-        ViewerRuntimeLiveServerConfig::new(WorldScenario::Minimal)
-            .with_decision_mode(ViewerLiveDecisionMode::Llm),
-    )
-    .expect("runtime server");
-    let agent_id = server
-        .world
-        .state()
-        .agents
-        .keys()
-        .next()
-        .cloned()
-        .expect("seed agent");
-    let (public_key, private_key) = test_signer(22);
-    let first = signed_agent_chat_request(
-        crate::viewer::AgentChatRequest {
-            agent_id: agent_id.clone(),
-            player_id: Some("player-a".to_string()),
-            public_key: None,
-            auth: None,
-            message: "hello".to_string(),
-            intent_tick: Some(10),
-            intent_seq: Some(6),
-        },
-        6,
-        public_key.as_str(),
-        private_key.as_str(),
-    );
-    let register_ack = register_runtime_session(
-        &mut server,
-        "player-a",
-        Some(agent_id.as_str()),
-        5,
-        public_key.as_str(),
-        private_key.as_str(),
-    );
-    assert_eq!(
-        register_ack.status,
-        AuthoritativeRecoveryStatus::SessionRegistered
-    );
-    server
-        .handle_agent_chat(first)
-        .expect("first request accepted");
-
-    let conflict = signed_agent_chat_request(
-        crate::viewer::AgentChatRequest {
-            agent_id,
-            player_id: Some("player-a".to_string()),
-            public_key: None,
-            auth: None,
-            message: "changed".to_string(),
-            intent_tick: Some(10),
-            intent_seq: Some(6),
-        },
-        6,
-        public_key.as_str(),
-        private_key.as_str(),
-    );
-    let err = server
-        .handle_agent_chat(conflict)
-        .expect_err("same seq with different payload must fail");
-    assert_eq!(err.code, "intent_seq_conflict");
-}
-
-#[test]
-fn runtime_agent_chat_rejects_intent_seq_nonce_mismatch() {
-    let mut server = ViewerRuntimeLiveServer::new(
-        ViewerRuntimeLiveServerConfig::new(WorldScenario::Minimal)
-            .with_decision_mode(ViewerLiveDecisionMode::Llm),
-    )
-    .expect("runtime server");
-    let agent_id = server
-        .world
-        .state()
-        .agents
-        .keys()
-        .next()
-        .cloned()
-        .expect("seed agent");
-    let (public_key, private_key) = test_signer(23);
-    let request = signed_agent_chat_request(
-        crate::viewer::AgentChatRequest {
-            agent_id: agent_id.clone(),
-            player_id: Some("player-a".to_string()),
-            public_key: None,
-            auth: None,
-            message: "hello".to_string(),
-            intent_tick: Some(3),
-            intent_seq: Some(8),
-        },
-        9,
-        public_key.as_str(),
-        private_key.as_str(),
-    );
-    let register_ack = register_runtime_session(
-        &mut server,
-        "player-a",
-        Some(agent_id.as_str()),
-        8,
-        public_key.as_str(),
-        private_key.as_str(),
-    );
-    assert_eq!(
-        register_ack.status,
-        AuthoritativeRecoveryStatus::SessionRegistered
-    );
-    let err = server
-        .handle_agent_chat(request)
-        .expect_err("intent seq mismatch should fail");
-    assert_eq!(err.code, "intent_seq_invalid");
-}
-
-#[test]
-fn runtime_authoritative_recovery_rotate_and_revoke_session_enforced_for_agent_chat() {
-    let _guard = lock_test_llm_env();
-    let mut server = ViewerRuntimeLiveServer::new(
-        ViewerRuntimeLiveServerConfig::new(WorldScenario::Minimal)
-            .with_decision_mode(ViewerLiveDecisionMode::Llm),
-    )
-    .expect("runtime server");
-    let agent_id = server
-        .world
-        .state()
-        .agents
-        .keys()
-        .next()
-        .cloned()
-        .expect("seed agent");
-    let (public_key_v1, private_key_v1) = test_signer(31);
-    let (public_key_v2, private_key_v2) = test_signer(32);
-
-    let first_request = signed_agent_chat_request(
-        crate::viewer::AgentChatRequest {
-            agent_id: agent_id.clone(),
-            player_id: Some("player-a".to_string()),
-            public_key: None,
-            auth: None,
-            message: "hello".to_string(),
-            intent_tick: Some(1),
-            intent_seq: Some(2),
-        },
-        2,
-        public_key_v1.as_str(),
-        private_key_v1.as_str(),
-    );
-    let register_ack = register_runtime_session(
-        &mut server,
-        "player-a",
-        Some(agent_id.as_str()),
-        1,
-        public_key_v1.as_str(),
-        private_key_v1.as_str(),
-    );
-    assert_eq!(
-        register_ack.status,
-        AuthoritativeRecoveryStatus::SessionRegistered
-    );
-    assert_eq!(register_ack.agent_id.as_deref(), Some(agent_id.as_str()));
-    let _ = server
-        .handle_agent_chat(first_request)
-        .expect("first key should be accepted");
-
-    let (rotate_ack, emit_snapshot_after_ack) = server
-        .handle_authoritative_recovery(AuthoritativeRecoveryCommand::RotateSession {
-            request: AuthoritativeSessionRotateRequest {
-                player_id: "player-a".to_string(),
-                old_session_pubkey: public_key_v1.clone(),
-                new_session_pubkey: public_key_v2.clone(),
-                rotate_reason: "security_rotation".to_string(),
-                rotated_by: Some("ops".to_string()),
-            },
-        })
-        .expect("rotate session");
-    assert!(!emit_snapshot_after_ack);
-    assert_eq!(
-        rotate_ack.status,
-        AuthoritativeRecoveryStatus::SessionRotated
-    );
-    assert_eq!(
-        rotate_ack.session_pubkey.as_deref(),
-        Some(public_key_v1.as_str())
-    );
-    assert_eq!(
-        rotate_ack.replaced_by_pubkey.as_deref(),
-        Some(public_key_v2.as_str())
-    );
-
-    let stale_request = signed_agent_chat_request(
-        crate::viewer::AgentChatRequest {
-            agent_id: agent_id.clone(),
-            player_id: Some("player-a".to_string()),
-            public_key: None,
-            auth: None,
-            message: "stale".to_string(),
-            intent_tick: Some(2),
-            intent_seq: Some(2),
-        },
-        2,
-        public_key_v1.as_str(),
-        private_key_v1.as_str(),
-    );
-    let stale_err = server
-        .handle_agent_chat(stale_request)
-        .expect_err("old key should be rejected after rotation");
-    assert_eq!(stale_err.code, "session_revoked");
-
-    let rotated_request = signed_agent_chat_request(
-        crate::viewer::AgentChatRequest {
-            agent_id: agent_id.clone(),
-            player_id: Some("player-a".to_string()),
-            public_key: None,
-            auth: None,
-            message: "rotated".to_string(),
-            intent_tick: Some(3),
-            intent_seq: Some(1),
-        },
-        1,
-        public_key_v2.as_str(),
-        private_key_v2.as_str(),
-    );
-    let _ = server
-        .handle_agent_chat(rotated_request)
-        .expect("new key should be accepted");
-
-    let (revoke_ack, emit_snapshot_after_ack) = server
-        .handle_authoritative_recovery(AuthoritativeRecoveryCommand::RevokeSession {
-            request: AuthoritativeSessionRevokeRequest {
-                player_id: "player-a".to_string(),
-                session_pubkey: Some(public_key_v2.clone()),
-                revoke_reason: "compromised".to_string(),
-                revoked_by: Some("ops".to_string()),
-            },
-        })
-        .expect("revoke session");
-    assert!(!emit_snapshot_after_ack);
-    assert_eq!(
-        revoke_ack.status,
-        AuthoritativeRecoveryStatus::SessionRevoked
-    );
-    assert_eq!(revoke_ack.revoke_reason.as_deref(), Some("compromised"));
-    assert_eq!(revoke_ack.revoked_by.as_deref(), Some("ops"));
-
-    let revoked_reconnect_err = server
-        .handle_authoritative_recovery(AuthoritativeRecoveryCommand::ReconnectSync {
-            request: AuthoritativeReconnectSyncRequest {
-                player_id: "player-a".to_string(),
-                session_pubkey: Some(public_key_v2.clone()),
-                last_known_log_cursor: None,
-                expected_reorg_epoch: None,
-            },
-        })
-        .expect_err("reconnect should surface revoke metadata");
-    assert_eq!(revoked_reconnect_err.code, "session_revoked");
-    assert_eq!(
-        revoked_reconnect_err.revoke_reason.as_deref(),
-        Some("compromised")
-    );
-    assert_eq!(revoked_reconnect_err.revoked_by.as_deref(), Some("ops"));
-
-    let revoked_register_request = signed_session_register_request(
-        crate::viewer::AuthoritativeSessionRegisterRequest {
-            player_id: "player-a".to_string(),
-            public_key: None,
-            auth: None,
-            requested_agent_id: Some(agent_id.clone()),
-            force_rebind: false,
-        },
-        5,
-        public_key_v2.as_str(),
-        private_key_v2.as_str(),
-    );
-    let revoked_register_err = server
-        .handle_authoritative_recovery(AuthoritativeRecoveryCommand::RegisterSession {
-            request: revoked_register_request,
-        })
-        .expect_err("register should surface revoke metadata");
-    assert_eq!(revoked_register_err.code, "session_revoked");
-    assert_eq!(
-        revoked_register_err.revoke_reason.as_deref(),
-        Some("compromised")
-    );
-    assert_eq!(revoked_register_err.revoked_by.as_deref(), Some("ops"));
-
-    let revoked_request = signed_agent_chat_request(
-        crate::viewer::AgentChatRequest {
-            agent_id,
-            player_id: Some("player-a".to_string()),
-            public_key: None,
-            auth: None,
-            message: "revoked".to_string(),
-            intent_tick: Some(4),
-            intent_seq: Some(2),
-        },
-        2,
-        public_key_v2.as_str(),
-        private_key_v2.as_str(),
-    );
-    let revoked_err = server
-        .handle_agent_chat(revoked_request)
-        .expect_err("revoked key should be rejected");
-    assert_eq!(revoked_err.code, "session_revoked");
 }
